Apollo Server (GraphQL) Implementation
I have fixed the sever config as previously, the docker container was connected to the wrong database. I have adjusted the docker-compose.yml file to fix this.

To create the schema, resolvers and index file, I used chatGPT, but i am now going to go through it and learn how the program works.